Removed const modifier for color variable
The const modifier on the float4 color definition at line 295 seems to cause problems with Intel GPUs. I removed that modifier (I have NO idea if this is a bad thing to do) and the kernel now compiles on both Nvidia and Intel GPUs. This means that the ClearVolume FIJI plugin can now work on Intel GPUs, whereas it would crash otherwise.