Update to ClearGL 2.2.11 and JOGL 2.5.0
I was unable to make the Gradle build system work, so I added a
Maven POM. (The only code that actually gets deployed is Java code,
so using Maven with the pom-scijava BOM works out of the box.)

Unfortunately, when I run this with the latest imglib2-clearvolume
plugin in an up-to-date Fiji, I see there are rotation bugs,
probably introduced by my hasty changes to ClearGL earlier today.
Also: ClearVolume does not work with Java 21.